# feat(snuba-init): Faster cli initialization