## Blast Radius

This affects every snuba deployment as it fundamentally changes how the snuba cli commands are invoked.

## Before State
Every CLI command would be imported no matter what cli command was called. This would add about two seconds to startup time. 

## After State 
The CLI still discovers the files in this directory as before but does not actually run their code until they are invoked. Thus `snuba api` would only ever import the `snuba/cli/api.py` module

## Testing Notes
I ran the `snuba <cmd>` for every subcommand in the CLI directory and made sure it still works. Also made sure to test the startup speed:

master:
```
❯ time snuba api --help
2022-10-25 09:24:06,784 Initializing Snuba...
2022-10-25 09:24:06,786 Loading module snuba.cli.admin
2022-10-25 09:24:06,789 Loading module snuba.cli.api
2022-10-25 09:24:06,886 Loading module snuba.cli.bootstrap
2022-10-25 09:24:07,346 Loading module snuba.cli.bulk_load
2022-10-25 09:24:07,961 Loading module snuba.cli.cleanup
2022-10-25 09:24:07,962 Loading module snuba.cli.config
2022-10-25 09:24:07,964 Loading module snuba.cli.consumer
2022-10-25 09:24:07,980 Loading module snuba.cli.devserver
2022-10-25 09:24:07,981 Loading module snuba.cli.entities
2022-10-25 09:24:08,335 Loading module snuba.cli.migrations
2022-10-25 09:24:08,337 Loading module snuba.cli.multistorage_consumer
2022-10-25 09:24:08,339 Loading module snuba.cli.offline_replacer
2022-10-25 09:24:08,346 Loading module snuba.cli.optimize
2022-10-25 09:24:08,352 Loading module snuba.cli.replacer
2022-10-25 09:24:08,354 Loading module snuba.cli.subscriptions_executor
2022-10-25 09:24:09,569 Loading module snuba.cli.subscriptions_scheduler
2022-10-25 09:24:09,576 Loading module snuba.cli.subscriptions_scheduler_executor
2022-10-25 09:24:09,579 Loading module snuba.cli.test_consumer
2022-10-25 09:24:09,581 Snuba initialization took 2.796583355s
Usage: snuba api [OPTIONS]

Options:
  --bind TEXT          Address to listen on.
  --debug
  --log-level TEXT     Logging level to use.
  --processes INTEGER
  --threads INTEGER
  --help               Show this message and exit.
snuba api --help  2.60s user 0.42s system 85% cpu 3.560 total
```

current branch:
```
❯ time snuba api --help
2022-10-25 09:31:26,456 Initializing Snuba...
2022-10-25 09:31:26,456 Snuba initialization took 0.0006065120000000035s
2022-10-25 09:31:26,457 Loading command api
Usage: snuba api [OPTIONS]

Options:
  --bind TEXT          Address to listen on.
  --debug
  --log-level TEXT     Logging level to use.
  --processes INTEGER
  --threads INTEGER
  --help               Show this message and exit.
snuba api --help  0.35s user 0.12s system 73% cpu 0.639 total
```